Use `X11` as fallback Hotkey implementation
`evdev` requires the user to be in the `input` group. This is the case
for some Linux distributions, but not all. We now try to identify if the
user is in that group and if that's not the case we use `X11` to grab
the hotkeys. This mostly matches our previous implementation, except
that it now handles the modifiers as well. Handling repeating hotkeys
does not seem to (at least easily) be possible as it's impossible to
differentiate a repeating hotkey from a newly pressed one. We usually
would then store the state of the key to see if it ever got released,
but here we also get released events even though the key never actually
got released. Since X11 is mostly a fallback implementation, this is
probably not too bad.
