+++ crates/livesplit-hotkey/Cargo.toml
@@ -22,7 +22,9 @@ objc = "0.2.7"
 [target.'cfg(target_os = "linux")'.dependencies]
 evdev = { version = "=0.11.4", optional = true }
 mio = { version = "0.8.0", default-features = false, features = ["os-ext", "os-poll"], optional = true }
+nix = { version = "0.25.0", features = ["user"], optional = true }
 promising-future = { version = "0.2.4", optional = true }
+x11-dl = { version = "2.20.0", optional = true }
 
 [target.'cfg(all(target_arch = "wasm32", target_os = "unknown"))'.dependencies]
 wasm-bindgen = { version = "0.2.54", optional = true }
@@ -36,5 +38,5 @@ bitflags = { version = "1.2.1" }
 
 [features]
 default = ["std"]
-std = ["snafu/std", "serde/std", "evdev", "mio", "promising-future", "winapi"]
+std = ["snafu/std", "serde/std", "evdev", "mio", "nix", "promising-future", "winapi", "x11-dl"]
 wasm-web = ["wasm-bindgen", "web-sys"]
